requirements.txt
pandas
numpy
torch
transformers
faiss-cpu
seaborn
matplotlib
sentencepiece
tiktoken 
tqdm